fix: use f64 for histogram priority filter to match block priority type
VisibilityHistogramQuery had priority_min/priority_max as Option<i32>,
but block priorities are f64. Sending a float value like 3.5 from the
priority slider caused serde to fail deserializing the query params,
producing a 422 error and leaving the histogram stuck in an error state
even after resetting the filter.

Changes:
- VisibilityHistogramQuery.priority_min/max: Option<i32> -> Option<f64>
- fetch_blocks_for_histogram trait + postgres/local implementations: same
- BlockHistogramData.priority: i32 -> f64 (removes precision-losing cast)
- compute_visibility_histogram_rust: priority_min/max: Option<i32> -> f64
- Remove redundant as-f64 casts in block_matches_visibility_histogram_query
- useVisibilityHistogram: add placeholderData: keepPreviousData so the
  previous histogram remains visible while a new filter query is loading,
  preventing the misleading 'No visibility data' flash
- Update all affected tests

+++ backend/src/db/models.rs
@@ -14,7 +14,7 @@ pub use crate::routes::insights::{
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct BlockHistogramData {
     pub scheduling_block_id: i64,
-    pub priority: i32,
+    pub priority: f64,
     pub visibility_periods: Option<Vec<Period>>,
 }
