+++ backend/src/db/repository/visualization.rs
@@ -52,8 +52,8 @@ pub trait VisualizationRepository: Send + Sync {
     async fn fetch_blocks_for_histogram(
         &self,
         schedule_id: crate::api::ScheduleId,
-        priority_min: Option<i32>,
-        priority_max: Option<i32>,
+        priority_min: Option<f64>,
+        priority_max: Option<f64>,
         block_ids: Option<Vec<i64>>,
     ) -> RepositoryResult<Vec<crate::db::models::BlockHistogramData>>;
